feat: integrated an svg component
