+++ src/pages/navigation/navigation.component.jsx
@@ -1,0 +1,25 @@
+import { Outlet, Link } from "react-router-dom";
+
+import { ReactComponent as CrwnLogo } from "../../assets/crown.svg";
+
+console.log(CrwnLogo);
+
+const Navigation = () => {
+  return (
+    <>
+      <div className="navigation">
+        <Link className="logo-container" to="/">
+          <CrwnLogo />
+        </Link>
+        <div className="nav-links-container">
+          <Link className="nav-link" to="/shop">
+            SHOP
+          </Link>
+        </div>
+      </div>
+      <Outlet />
+    </>
+  );
+};
+
+export default Navigation;
